+++ Project/Engine/CRenderMgr.cpp
@@ -119,16 +119,16 @@ void CRenderMgr::CopyRenderTarget()
 void CRenderMgr::UpdateData()
 {
     // GlobalData 에 광원 개수정보 세팅
-    GlobalData.Light2DCount = m_vecLight2D.size();
+    GlobalData.Light2DCount = (UINT)m_vecLight2D.size();
 
     // 구조화버퍼의 크기가 모자라면 더 크게 새로 만든다.
-    if (m_Light2DBuffer->GetElementCount() < m_vecLight2D.size())
+    if (m_Light2DBuffer->GetElementCount() < (UINT)m_vecLight2D.size())
     {
-        m_Light2DBuffer->Create(sizeof(tLightInfo), m_vecLight2D.size(), SB_TYPE::READ_ONLY, true);
+        m_Light2DBuffer->Create(sizeof(tLightInfo), (UINT)m_vecLight2D.size(), SB_TYPE::READ_ONLY, true);
     }
 
     // 구조화버퍼로 광원 데이터를 옮긴다.
-    m_Light2DBuffer->SetData(m_vecLight2D.data(), sizeof(tLightInfo) * m_vecLight2D.size());
+    m_Light2DBuffer->SetData(m_vecLight2D.data(), sizeof(tLightInfo) * (UINT)m_vecLight2D.size());
     m_Light2DBuffer->UpdateData(12, PIPELINE_STAGE::PS_PIXEL);
